added log

+++ src/q/rest/quotation/operation/sockets/CustomerNotificationEndPoint.java
@@ -32,7 +32,7 @@ public class CustomerNotificationEndPoint {
 
 
     @OnOpen
-    public void onOpen(Session session, @PathParam("customerId") Integer customerId, @PathParam("token") String token) throws IOException {
+    public void onOpen(Session session, @PathParam("customerId") Long customerId, @PathParam("token") String token) throws IOException {
         System.out.println("openning customer id " + customerId);
         this.session = session;
         this.customerId = customerId.longValue();
